+++ src/Command/TemplateWriteCommand.php
@@ -25,6 +25,8 @@ use Webwerkwien\ContaoAiCoreBundle\Service\SystemLog;
 #[AsCommand(name: 'contao:template:write', description: 'Write a Twig template to the correct path under templates/')]
 class TemplateWriteCommand extends Command
 {
+    use JsonErrorBoundary;
+
     use OperatorOptionTrait;
 
     private LoggerInterface $logger;
@@ -59,6 +61,11 @@ class TemplateWriteCommand extends Command
     }
 
     protected function execute(InputInterface $input, OutputInterface $output): int
+    {
+        return $this->guarded($output, fn (): int => $this->doExecute($input, $output));
+    }
+
+    protected function doExecute(InputInterface $input, OutputInterface $output): int
     {
         $mode   = $input->getOption('mode');
         $base   = $input->getOption('base');